Add ability to add expenses
